// reintegros-frontend/components/solicitudes/prestaciones/types.ts
export type FormDataPrestaciones = {
  //entity fields here
  nomenclador: any;
  valorIprossNomenclador: number;
  coseguroNomenclador: number;
  montoReintegrar: number;
  cobertura: string;
  modalidad: string;
  estadoActual: string;
  cantidad: number;
  montoTotal: number;
  // unidad: string;
};